扩展温度类型 int8_t -> int16_t, 避免 ±127°C 截断
motorProValue.board_temp / motor_temp 之前是 int8_t, 在故障/失控
温升场景下 (>127°C) 会绕回到负数, 导致过温保护误判. 阈值 TemBorad=100
TemMortor=120 在 int8 上限附近, 留的余量太薄.

改动:
- ifly_fault.h: motorProValue 字段 + getter + NTC 转换函数全部 int16_t
- ifly_fault.c: NTC 函数体本地变量同步; boradTempProFunc / motorTempProFunc
  内 (int8_t)Threshold.Tem* cast 改为 (int16_t), 阈值本身 uint16_t
- ifly_fault_api.{c,h}: 三个 setter stub 入参类型同步
- can_wly.c: 状态帧 / 扩展状态帧三处温度比较 cast 同步, 顺手去掉
  temp_motor/temp_mos 多余 (int16_t) 中间 cast
- foc_bsp.c: logid 151 局部 t_board_c / t_motor_c 类型同步

NTC 表本身的有效区间没动 (电机端查表 -20~+200°C 饱和; 板温 Beta 公式
0~150°C 精度可信). 只是把承载温度值的容器扩到 int16_t, 让阈值/比较/上报
全程不再受 int8 截断影响.

编译: 0 Error 0 Warning, Code 105236 -> 105244 (+8 字节)

+++ Core/Src/can_wly.c
@@ -211,8 +211,8 @@ static void pack_status_frame(uint8_t *d) {
 
     /* WARN: Bit0=MOS过温警告(90°C), Bit1=电机过温警告 */
     uint8_t  warn = 0;
-    if (motorProValue.board_temp >= (int8_t)Threshold.TemBoradWarn) warn |= 0x01;
-    if (motorProValue.motor_temp >= (int8_t)Threshold.TemMortorWarn) warn |= 0x02;
+    if (motorProValue.board_temp >= (int16_t)Threshold.TemBoradWarn) warn |= 0x01;
+    if (motorProValue.motor_temp >= (int16_t)Threshold.TemMortorWarn) warn |= 0x02;
 
     d[0] = p_int & 0xFF;
     d[1] = (p_int >> 8) & 0xFF;
@@ -269,9 +269,9 @@ static void send_ext_status_frame(void) {
     /* 位置: real_position_out (1°/1024) → 0.001° */
     int32_t p_int = (int32_t)(((int64_t)controller_eyou.real_position_out * 1000) / 1024);
 
-    /* 温度: int8 °C → int16 0.1°C (有符号, 负温度也正确) */
-    int16_t temp_motor = (int16_t)motorProValue.motor_temp * 10;
-    int16_t temp_mos   = (int16_t)motorProValue.board_temp * 10;
+    /* 温度: int16 °C → int16 0.1°C (有符号, 负温度也正确) */
+    int16_t temp_motor = motorProValue.motor_temp * 10;
+    int16_t temp_mos   = motorProValue.board_temp * 10;
 
     d[0] = i_rms_100 & 0xFF;
     d[1] = (i_rms_100 >> 8) & 0xFF;
@@ -289,7 +289,7 @@ static void send_ext_status_frame(void) {
     uint8_t sta = 0;
     if (controller_eyou.foc_run) sta |= 0x01;
     if (controller_eyou.ServoErrFlag.All_Flag) sta |= 0x02;
-    if (motorProValue.board_temp >= (int8_t)Threshold.TemBoradWarn) sta |= 0x04;
+    if (motorProValue.board_temp >= (int16_t)Threshold.TemBoradWarn) sta |= 0x04;
     if (controller_eyou.ServoState.Bit.PositionArrivedFlag) sta |= 0x08;
     d[12] = sta;
